Moved e2e tests to own folder
